feat(chat): implement chat-specific routes with sidebar and improved message alignment
Refactor generate route to support nested chat ID routes, add sidebar with new chat button, and fix message entry styling for proper user/agent alignment. This enables multiple concurrent chats with unique IDs.
