+++ app/lib/my-components/MessageEntry.tsx
@@ -9,13 +9,13 @@ interface Message {
 export const MessageEntry: React.FC<{ message: Message }> = ({ message }) => {
   return (
         <p className={clsx({
-            "bg-gray-100 text-gray-800": message.role === "user",
-            "bg-gray-800 text-gray-100": message.role === "agent",
+            "bg-gray-100 text-gray-800 self-end": message.role === "user",
+            "bg-gray-800 text-gray-100 self-start": message.role === "agent",
         }, 
         "max-w-2/3",
         "rounded-md",
         "p-4",
-        "odd:self-end even:self-start",
+        // "odd:self-end even:self-start",
         "text-left",
         "wrap-break-word"
         )}>
